move database calls to data-access library
