+++ apps/api-graphql/src/graphql/schema/passage/passage.mutation.ts
@@ -1,113 +1,25 @@
 import type { GraphQLContext } from '../../context';
 import {
-  ANNOTATIONS_TO_IGNORE,
-  DataClient,
   Passage,
   Annotation,
   findLiteralOccurrences,
+  fetchReplaceAnnotations,
+  fetchReplaceRows,
+  hasPermission,
+  persistReplaceChanges,
   replacePassageText,
-  passagesToDTO,
-  passagesToRowDTO,
+  savePassagesWithDeletions,
 } from '@eightyfourthousand/data-access';
 import {
   buildAnnotationsByPassageUuid,
   buildReplacePassage,
   buildReplaceResponsePassages,
-  fetchReplaceAnnotations,
-  fetchReplaceRows,
   findNextReplaceCursor,
-  persistReplaceChanges,
   replaceFailure,
   replaceSuccess,
-  type ReplacePassageRow,
   type ReplaceResult,
   validateReplaceArgs,
 } from './passage.replace';
-
-const PAGE_SIZE = 500;
-
-async function normalizePassageLabelsAfter(
-  supabase: DataClient,
-  workUuid: string,
-  fromSort: number,
-  fromLabel: string,
-  delta: number, // +1 for insert, -1 for delete
-): Promise<void> {
-  const parts = fromLabel.split('.');
-  const depth = parts.length;
-  const prefix = depth > 1 ? parts.slice(0, -1).join('.') + '.' : '';
-  let nextInt = parseInt(parts[depth - 1], 10) + Math.max(delta, 0);
-
-  let lastSort = fromSort;
-  let done = false;
-
-  while (!done) {
-    const { data, error } = await supabase
-      .from('passages')
-      .select('uuid, label, sort')
-      .eq('work_uuid', workUuid)
-      .gt('sort', lastSort)
-      .order('sort', { ascending: true })
-      .limit(PAGE_SIZE);
-
-    if (error || !data || data.length === 0) break;
-
-    const labelUpdates: { uuid: string; label: string }[] = [];
-    const prefixRenames: { oldPrefix: string; newPrefix: string }[] = [];
-
-    for (const row of data) {
-      const rowParts = (row.label ?? '').split('.');
-
-      if (rowParts.length < depth) {
-        done = true;
-        break;
-      } // shallower → stop
-      if (!row.label.startsWith(prefix)) {
-        done = true;
-        break;
-      } // different prefix → stop
-      if (rowParts.length > depth) continue; // deeper → skip (handled by prefix rename)
-
-      const expectedLabel = prefix + nextInt;
-      if (row.label === expectedLabel) {
-        done = true;
-        break;
-      } // already contiguous → stop
-
-      labelUpdates.push({ uuid: row.uuid, label: expectedLabel });
-      prefixRenames.push({
-        oldPrefix: row.label + '.',
-        newPrefix: expectedLabel + '.',
-      });
-      nextInt++;
-    }
-
-    if (labelUpdates.length > 0) {
-      const { error: upsertError } = await supabase
-        .from('passages')
-        .upsert(labelUpdates, { onConflict: 'uuid' });
-      if (upsertError)
-        console.error('Error normalizing passage labels:', upsertError);
-
-      // Fire child prefix renames after same-depth labels are committed
-      for (const { oldPrefix, newPrefix } of prefixRenames) {
-        const { error: prefixError } = await supabase.rpc(
-          'rename_passage_label_prefix',
-          {
-            p_work_uuid: workUuid,
-            p_old_prefix: oldPrefix,
-            p_new_prefix: newPrefix,
-          },
-        );
-        if (prefixError)
-          console.error('Error renaming passage label prefix:', prefixError);
-      }
-    }
-
-    if (done || data.length < PAGE_SIZE) break;
-    lastSort = data[data.length - 1].sort;
-  }
-}
 
 interface PassageInput {
   uuid: string;
@@ -137,14 +49,12 @@ const requireEditorEditPermission = async (ctx: GraphQLContext) => {
     };
   }
 
-  const { data: hasPermission, error: permError } = await ctx.supabase.rpc(
-    'authorize',
-    {
-      requested_permission: 'editor.edit',
-    },
-  );
+  const permitted = await hasPermission({
+    client: ctx.supabase,
+    permission: 'editor.edit',
+  });
 
-  if (permError || !hasPermission) {
+  if (!permitted) {
     return {
       ok: false as const,
       error: 'Permission denied: editor.edit required',
@@ -185,167 +95,11 @@ export const savePassagesMutation = async (
       annotations: input.annotations as Annotation[],
     }));
 
-    // Phase A — Detect new passages
-    const inputUuids = passages.map((p) => p.uuid);
-    const { data: existingRows } = await ctx.supabase
-      .from('passages')
-      .select('uuid')
-      .in('uuid', inputUuids);
-    const existingUuidSet = new Set((existingRows ?? []).map((r) => r.uuid));
-    const newPassages = passages.filter((p) => !existingUuidSet.has(p.uuid));
-
-    // Phase B — Sort shift for new passages (descending sort order to avoid conflicts)
-    const sortedNewPassages = [...newPassages].sort((a, b) => b.sort - a.sort);
-    for (const p of sortedNewPassages) {
-      const { error: shiftError } = await ctx.supabase.rpc(
-        'shift_passage_sorts',
-        {
-          p_work_uuid: p.workUuid,
-          p_from_sort: p.sort,
-          p_delta: 1,
-        },
-      );
-      if (shiftError) {
-        console.error('Error shifting passage sorts:', shiftError);
-      }
-    }
-
-    // Phase C — Normalize labels for new passages
-    for (const p of sortedNewPassages) {
-      await normalizePassageLabelsAfter(
-        ctx.supabase,
-        p.workUuid,
-        p.sort,
-        p.label,
-        1,
-      );
-    }
-
-    // Phase D — Handle deletions
-    let deletedCount = 0;
-    if (args.deletedUuids && args.deletedUuids.length > 0) {
-      // Fetch deleted passage info
-      const { data: deletedPassages } = await ctx.supabase
-        .from('passages')
-        .select('uuid, sort, label, work_uuid')
-        .in('uuid', args.deletedUuids);
-
-      if (deletedPassages && deletedPassages.length > 0) {
-        // Adjust labels for deleted passages (descending sort order)
-        const sortedDeleted = [...deletedPassages].sort(
-          (a, b) => b.sort - a.sort,
-        );
-        for (const dp of sortedDeleted) {
-          await normalizePassageLabelsAfter(
-            ctx.supabase,
-            dp.work_uuid,
-            dp.sort,
-            dp.label,
-            -1,
-          );
-        }
-
-        // Delete annotations first (passage_annotations has ON DELETE RESTRICT)
-        const { error: deleteAnnotationsError } = await ctx.supabase
-          .from('passage_annotations')
-          .delete()
-          .in('passage_uuid', args.deletedUuids);
-
-        if (deleteAnnotationsError) {
-          console.error(
-            'Error deleting annotations for deleted passages:',
-            deleteAnnotationsError,
-          );
-        }
-
-        // Delete passages
-        const { error: deletePassagesError } = await ctx.supabase
-          .from('passages')
-          .delete()
-          .in('uuid', args.deletedUuids);
-
-        if (deletePassagesError) {
-          console.error('Error deleting passages:', deletePassagesError);
-          return {
-            success: false,
-            savedCount: 0,
-            error: `Failed to delete passages: ${deletePassagesError.message}`,
-          };
-        }
-
-        deletedCount = deletedPassages.length;
-      }
-    }
-
-    // Phase E — Existing upsert logic
-    // Convert to DTOs
-    const dtos = passagesToDTO(passages);
-    const passageRowDtos = passagesToRowDTO(passages);
-    const passageUuids = passages.map((p) => p.uuid);
-    const annotations = dtos.flatMap((p) => p.annotations || []);
-
-    // Get existing annotations to determine deletions
-    const { data: existingAnnotations } = await ctx.supabase
-      .from('passage_annotations')
-      .select('uuid')
-      .in('passage_uuid', passageUuids)
-      .not('type', 'in', `(${ANNOTATIONS_TO_IGNORE.join(',')})`);
-
-    const annotationsToDelete = existingAnnotations?.filter(
-      (ea) => !annotations.find((a) => a.uuid === ea.uuid),
-    );
-
-    // Upsert passages
-    const { error: passageError } = await ctx.supabase
-      .from('passages')
-      .upsert(passageRowDtos, { onConflict: 'uuid' });
-
-    if (passageError) {
-      console.error('Error saving passages:', passageError);
-      return {
-        success: false,
-        savedCount: 0,
-        error: `Failed to save passages: ${passageError.message}`,
-      };
-    }
-
-    // Upsert annotations
-    if (annotations.length > 0) {
-      const { error: annotationError } = await ctx.supabase
-        .from('passage_annotations')
-        .upsert(annotations, { onConflict: 'uuid' });
-
-      if (annotationError) {
-        console.error('Error saving annotations:', annotationError);
-        return {
-          success: false,
-          savedCount: passages.length,
-          error: `Passages saved but annotations failed: ${annotationError.message}`,
-        };
-      }
-    }
-
-    // Delete removed annotations
-    if (annotationsToDelete && annotationsToDelete.length > 0) {
-      const { error: deleteError } = await ctx.supabase
-        .from('passage_annotations')
-        .delete()
-        .in(
-          'uuid',
-          annotationsToDelete.map((a) => a.uuid),
-        );
-
-      if (deleteError) {
-        console.error('Error deleting annotations:', deleteError);
-        // Non-fatal, continue
-      }
-    }
-
-    return {
-      success: true,
-      savedCount: passages.length,
-      deletedCount,
-    };
+    return savePassagesWithDeletions({
+      client: ctx.supabase,
+      passages,
+      deletedUuids: args.deletedUuids,
+    });
   } catch (error) {
     console.error('Unexpected error saving passages:', error);
     return {
@@ -386,14 +140,20 @@ export const replaceMutation = async (
   }
 
   try {
-    const rowsResult = await fetchReplaceRows({ ctx, targetUuids });
+    const rowsResult = await fetchReplaceRows({
+      client: ctx.supabase,
+      targetUuids,
+    });
     if (!rowsResult.ok) {
-      return rowsResult.result;
+      return replaceFailure(rowsResult.error);
     }
 
-    const annotationsResult = await fetchReplaceAnnotations({ ctx, targetUuids });
+    const annotationsResult = await fetchReplaceAnnotations({
+      client: ctx.supabase,
+      targetUuids,
+    });
     if (!annotationsResult.ok) {
-      return annotationsResult.result;
+      return replaceFailure(annotationsResult.error);
     }
 
     const orderedRows = rowsResult.orderedRows;
@@ -497,12 +257,12 @@ export const replaceMutation = async (
     }
 
     const persistResult = await persistReplaceChanges({
-      ctx,
+      client: ctx.supabase,
       rawAnnotations,
       updatedPassages,
     });
     if (!persistResult.ok) {
-      return persistResult.result;
+      return replaceFailure(persistResult.error);
     }
 
     return replaceSuccess({
