+++ apps/api-graphql/src/graphql/schema/work/title.resolver.ts
@@ -1,6 +1,6 @@
 import type { GraphQLContext } from '../../context';
 import type { WorkParent } from './work.types';
-import type { TitleDTO } from '@eightyfourthousand/data-access';
+import { getWorkTitles } from '@eightyfourthousand/data-access';
 
 type TitleResult = {
   uuid: string;
@@ -14,22 +14,11 @@ export const titlesResolver = async (
   _args: Record<string, never>,
   ctx: GraphQLContext,
 ): Promise<TitleResult[]> => {
-  const { data, error } = await ctx.supabase.rpc('get_work_titles', {
-    work_uuid_input: parent.uuid,
-  });
-
-  if (error) {
-    console.error('Error fetching work titles:', error);
-    return [];
-  }
-
-  const titles = (data ?? []) as TitleDTO[];
-
+  const titles = await getWorkTitles({ client: ctx.supabase, uuid: parent.uuid });
   return titles.map((title) => ({
     uuid: title.uuid,
     content: title.title,
     language: title.language,
-    // Remove the 'eft:' prefix from the type
-    type: title.type?.replace('eft:', '') || 'mainTitle',
+    type: title.type,
   }));
 };
