+++ apps/api-graphql/src/graphql/schema/work/work-title.loader.ts
@@ -1,5 +1,8 @@
 import DataLoader from 'dataloader';
-import type { DataClient } from '@eightyfourthousand/data-access';
+import {
+  getWorkTitlesByUuids,
+  type DataClient,
+} from '@eightyfourthousand/data-access';
 
 /**
  * Creates a DataLoader for batch-fetching work titles by UUID.
@@ -7,23 +10,10 @@ import type { DataClient } from '@eightyfourthousand/data-access';
  */
 export function createWorkTitleLoader(supabase: DataClient) {
   return new DataLoader<string, string | null>(async (workUuids) => {
-    const { data, error } = await supabase
-      .from('works')
-      .select('uuid, title')
-      .in('uuid', workUuids as string[]);
-
-    if (error) {
-      console.error('Error batch loading work titles:', error);
-      return workUuids.map(() => null);
-    }
-
-    const titlesByUuid = new Map<string, string>();
-    for (const work of data ?? []) {
-      if (work.title) {
-        titlesByUuid.set(work.uuid, work.title);
-      }
-    }
-
+    const titlesByUuid = await getWorkTitlesByUuids({
+      client: supabase,
+      uuids: workUuids,
+    });
     return workUuids.map((uuid) => titlesByUuid.get(uuid) ?? null);
   });
 }
