+++ libs/data-access/src/lib/folio.ts
@@ -1,23 +1,26 @@
 'use server';
 
 import { createServerClient } from './client-ssr';
-import { TohokuCatalogEntry } from './types';
+import { DataClient, TohokuCatalogEntry } from './types';
 import { FolioDTO, folioFromDTO } from './types/folio';
 
-export const getFolios = async ({
-  uuid,
-  toh,
-  page = 0,
-  size = 10,
-}: {
+type GetWorkFoliosArgs = {
+  client: DataClient;
   uuid: string;
   toh: TohokuCatalogEntry;
   page?: number;
   size?: number;
-}) => {
+};
+
+export const getWorkFolios = async ({
+  client,
+  uuid,
+  toh,
+  page = 0,
+  size = 10,
+}: GetWorkFoliosArgs) => {
   const start = page * size;
   const end = start + size - 1;
-  const client = await createServerClient();
   const { data, error } = await client
     .from('tibetan_works_folios')
     .select(
@@ -30,6 +33,7 @@ export const getFolios = async ({
     )
     .eq('work_uuid', uuid)
     .eq('toh', toh)
+    .order('volume_number', { ascending: true })
     .order('folio_number', { ascending: true })
     .order('side', { ascending: true })
     .range(start, end);
@@ -40,4 +44,14 @@ export const getFolios = async ({
   }
 
   return data.map((dto) => folioFromDTO(dto as FolioDTO));
+};
+
+export const getFolios = async ({
+  uuid,
+  toh,
+  page = 0,
+  size = 10,
+}: Omit<GetWorkFoliosArgs, 'client'>) => {
+  const client = await createServerClient();
+  return getWorkFolios({ client, uuid, toh, page, size });
 };
