+++ libs/data-access/src/lib/publications.ts
@@ -1,6 +1,7 @@
 import {
   BodyItemType,
   DataClient,
+  Title,
   TitlesDTO,
   WorkDTO,
   titlesFromDTO,
@@ -13,6 +14,19 @@ import {
   PassagesPageAroundDTO,
   Work,
 } from './types';
+
+type WorksPageInfo = {
+  nextCursor: string | null;
+  prevCursor: string | null;
+  hasMoreAfter: boolean;
+  hasMoreBefore: boolean;
+};
+
+export type WorksPage = {
+  items: Work[];
+  pageInfo: WorksPageInfo;
+  totalCount: number;
+};
 
 export const getTranslationUuids = async ({
   client,
@@ -108,6 +122,54 @@ export const getTranslationTitles = async ({
   });
 
   return titlesFromDTO(data as TitlesDTO);
+};
+
+export const getWorkTitles = async ({
+  client,
+  uuid,
+}: {
+  client: DataClient;
+  uuid: string;
+}): Promise<Title[]> => {
+  const { data, error } = await client.rpc('get_work_titles', {
+    work_uuid_input: uuid,
+  });
+
+  if (error) {
+    console.error('Error fetching work titles:', error);
+    return [];
+  }
+
+  return titlesFromDTO(data as TitlesDTO);
+};
+
+export const getWorkTitlesByUuids = async ({
+  client,
+  uuids,
+}: {
+  client: DataClient;
+  uuids: readonly string[];
+}): Promise<Map<string, string>> => {
+  const titlesByUuid = new Map<string, string>();
+  if (uuids.length === 0) return titlesByUuid;
+
+  const { data, error } = await client
+    .from('works')
+    .select('uuid, title')
+    .in('uuid', uuids as string[]);
+
+  if (error) {
+    console.error('Error batch loading work titles:', error);
+    return titlesByUuid;
+  }
+
+  for (const work of data ?? []) {
+    if (work.title) {
+      titlesByUuid.set(work.uuid, work.title);
+    }
+  }
+
+  return titlesByUuid;
 };
 
 export const getTranslationMetadataByUuid = async ({
@@ -210,4 +272,85 @@ export const getTranslationsMetadata = async ({
 
   const dto = data as WorkDTO[];
   return dto?.map((work) => workFromDTO(work as WorkDTO)) || [];
+};
+
+export const getWorksPage = async ({
+  client,
+  cursor,
+  limit = 50,
+  maxPages,
+}: {
+  client: DataClient;
+  cursor?: string;
+  limit?: number;
+  maxPages?: number;
+}): Promise<WorksPage> => {
+  const clampedLimit = Math.min(limit, 200);
+
+  let query = client
+    .from('works')
+    .select(
+      `
+      uuid,
+      title,
+      description,
+      tohs:work_toh!inner(toh:toh_clean),
+      publicationDate,
+      publicationVersion,
+      pages:source_pages,
+      restriction,
+      breadcrumb
+    `,
+      { count: 'exact' },
+    )
+    .not('toh', 'like', 'toh00%')
+    .order('title', { ascending: true })
+    .limit(clampedLimit + 1);
+
+  if (maxPages) {
+    query = query.lt('source_pages', maxPages);
+  }
+
+  if (cursor) {
+    const { data: cursorWork } = await client
+      .from('works')
+      .select('title')
+      .eq('uuid', cursor)
+      .single();
+
+    if (cursorWork) {
+      query = query.gt('title', cursorWork.title);
+    }
+  }
+
+  const { data, error, count } = await query;
+
+  if (error) {
+    console.error('Error fetching works:', error);
+    return {
+      items: [],
+      pageInfo: {
+        nextCursor: null,
+        prevCursor: null,
+        hasMoreAfter: false,
+        hasMoreBefore: false,
+      },
+      totalCount: 0,
+    };
+  }
+
+  const hasMoreAfter = (data ?? []).length > clampedLimit;
+  const items = hasMoreAfter ? (data ?? []).slice(0, clampedLimit) : data ?? [];
+  const works = items.map((dto) => workFromDTO(dto as WorkDTO));
+
+  return {
+    items: works,
+    pageInfo: {
+      nextCursor: hasMoreAfter ? works[works.length - 1]?.uuid ?? null : null,
+      prevCursor: cursor ?? null,
+      hasMoreAfter,
+      hasMoreBefore: Boolean(cursor),
+    },
+    totalCount: count ?? 0,
+  };
 };
